"Some more changes, a bugfix, and a new visual theme in beta."
I tagged RhythmLunatic in \BGAnimations\ScreenSelectMusic overlay\PerPlayer\PaneDisplay.lua for help with a working but very annoying math on a nil value instance. I wasn't aware that False would be an output of line 276 upon every time I switch a song.

A new visual theme, Simply Particles, is now in beta. I made the music to go along with it (\Sounds\_common menu music\aesthetical (loop).ogg) by editing a song block I found on an old disc for an instant music solution set of software that I had for many years. If Dan puts me on the line for such simplicity (pun kind of intended here), I would be amazed, but even then, I need to finish it, including redoing the icon and the title menu text.

A few of the other changes in this commit include a port of a feature from that you-know-what build for combo break flashes instead of just Miss. Sadly, the vibration doesn't happen anymore, as I was thinking of reapplying this, but lost the idea. I'm also not sure if doing this in other modes like ECFA or Stomp works anymore. I'll find a way eventually.

Another feature I thought about was putting in FC text for an FC. I also made certain colors relavant to the theme and added colorization for labels and numbers on evaluation pane 1 and labels on pane 2 for Competitive and Casual.

Co-Authored-By: Rhythm Lunatic <kirbyrulez@att.net>