Syncing is back!
Earlier, this version of SL had no gameplay syncing options. In such case, I brought it back, and better than ever, with custom music!

However, it only works in dance mode. Damn the .ssc file that was in there at first for syncing to steps and not song for the other modes. Hate when that happens, especially since I don't play around with that myself.