Hello, new intro, new sounds! Also, PARTICLES DAMMIT!
The Particles background at first looked to have half a particle at the bottom. That's fixed now, so enjoy the peace for now, those with OCD!

As I have done with a few sounds already, I made more sounds for the theme. That's right, made! Gotta love experimentation!

A new intro has finally been done in time for a newer theme I am a side developer of called Doki Doki Dancing All Night! I hope it makes it in, as I helped Gio come up with a few Easter egg ideas. We have yet to figure out how to put them in, but the mockups he's been sending me have been on the road to a very great story! Check out the "pre-pre-alpha" as we have called it here: https://github.com/RadicalRiko69/DDDAN/

As for updates with this theme, they will be more and more sparse as I have so many things coming up at once, including, but not limited to, school, more tournaments, and simfile development spikes I might sneak in.

Three evaluation Easter eggs for this theme have yet to be decided, but I don't need ideas. I'll figure it out myself, give me time... or is that another hint?